test: use systemd-run for running podman as admin user
To eventually support VSOCK for running commands in the virtual machine
we can no longer use SSHConnection to establish an administrator ssh
session. Additionally SSHConnection can be refactored without breaking
our tests nor do we rely on `m.ssh_port` or other machine properties.

There are two separate methods required to run commands, one
`run_as_admin` which returns the exit code and waits on the command to
complete. For `podman run -d` we need the command to "detach" and for
that we have to start the command as an unit and inspect if it runs
successfully ourself.